Rename County.cs to county.cs
For some reason the rename didn't happen when I pushed it. Also, I should mention that there's a district and scenario class now. The scenario class is basically just an array of districts, and the district class is just a number (the name of the district) and an array of counties.